Fix the Typo parttern instead of pattern
